% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncObject.r
\name{AncObject}
\alias{AncObject}
\title{Function to sort data according to ie output (following the ape tree structure)}
\usage{
AncObject(tree, data_extants, data_nodes)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{data_extants}{named vector of tip values}

\item{data_nodes}{named vector of ancestral states}
}
\value{
object with extant and nodal values sorted according to the ape tree structure
}
\description{
This function sorts extant and nodal values according the ape tree structure
}

