% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/EvoMorphoSpace.r
\name{EvoMorphoSpace}
\alias{EvoMorphoSpace}
\title{Evolutionary morphospace}
\usage{
EvoMorphoSpace(tree,X,Y,sequence,colors)
}
\arguments{
\item{tree}{object of class "phylo"}
\item{X}{an object created by "ieObject" for the X variable}
\item{Y}{an object created by "ieObject" for the Y variable}
\item{sequence}{vector of units of time for which the morphospace is to be computed}
\item{colors}{vector of colors to be assigned to each lineage}
}
\value{
a series of jpg files with the morphospace for each element in sequence
}
\description{
This function calculates ancestral values at a particular snapshot in evolutionare time
}
\author{Jeroen Smaers}

