\name{getEdges}
\alias{getEdges}
\title{Selecting edges descendant to a node}
\usage{
getEdges(tree,node)
}
\arguments{
\item{tree}{object of class "phylo"}
\item{node}{node number}
}
\value{
vector with edge labels
}
\description{
Function to locate all edges descendant to a node
}
\author{Jeroen Smaers}

