% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNodes.r
\name{getNodes}
\alias{getNodes}
\title{List all nodes ancestral to another node}
\usage{
getNodes(tree, node)
}
\arguments{
\item{tree}{an object of class 'phylo'.}

\item{node}{number of interest.}
}
\value{
vector of node numbers.
}
\description{
List all nodes ancestral to another node
}

