% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTips.r
\name{getTips}
\alias{getTips}
\title{List all tips descendant from a node}
\usage{
getTips(tree, node)
}
\arguments{
\item{tree}{an object of class 'phylo'.}

\item{node}{number of interest.}
}
\value{
vector of tip numbers.
}
\description{
List all tips descendant from a node
}
