% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.ci.r
\name{gls.ci}
\alias{gls.ci}
\title{GLS confidence intervals}
\usage{
gls.ci(Y, X, Sigma)
}
\arguments{
\item{Y}{Dependent variable.}

\item{X}{Independent variable.}

\item{Sigma}{Matrix describing the covariance among data points. For the phylogenetic regression, this is based on a phylogeny (vcv(tree)).}
}
\value{
GLS results.

Confidence intervals of the estimates ('CI').

Confidence intervals of unobserved values of Y ('CI.plot').
}
\description{
Computes confidence intervals for a GLS regression. When the variance covariance matrix is based on a phylogeny, this computes the confidence intervals for a phylogenetic regression (PGLS).
}
\examples{
see smaerslab.com/pGLSintervals/ 
}
\references{
Smaers & Rohlf (2016) Testing species' deviations from allometric preductions using the phylogenetic regression. Evolution. 70 (5): 1145-1149.
}
