% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.pi.r
\name{gls.pi}
\alias{gls.pi}
\title{GLS prediction intervals}
\usage{
gls.pi(Y, X, Sigma, k)
}
\arguments{
\item{Y}{Dependent variable.}

\item{X}{Independent variable.}

\item{Sigma}{Matrix describing the covariance among data points. For the phylogenetic regression, this is based on a phylogeny (vcv(tree)).}

\item{k}{The sample size of the estimated future points.}
}
\value{
GLS results.

Prediction intervals of the estimates ('PI').

Prediction intervals of unobserved values of Y ('PI.plot').
}
\description{
Computes prediction intervals for a GLS regression. When the variance covariance matrix is based on a phylogeny, this computes the prediction intervals for a phylogenetic regression (PGLS).
}
\examples{
see smaerslab.com/pGLSintervals/ 
}
\references{
Smaers & Rohlf (2016) Testing species' deviations from allometric preductions using the phylogenetic regression. Evolution. 70 (5): 1145-1149.
}
