% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ie.r
\name{ie}
\alias{ie}
\title{Function for the method of Independent Evolution}
\usage{
ie(data, tree)
}
\arguments{
\item{data}{numeric vector with elements in the same order as tiplabels in the tree}
\item{tree}{object of class "phylo"}
}
\value{
dataframe with rates for all branches and ancestral states for all nodes in the tree
}
\description{
This function allows computing rates and ancestral states using the method of Independent Evolution
}
\author{Jeroen Smaers}

