% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/iePlot.r
\name{iePlot}
\alias{iePlot}
\title{Function for mapping ie results onto a phylogenetic tree}
\usage{
iePlot(ieObject, tree, branchWidth, branchCutoff, labelSize, scaling)
}
\arguments{
\item{ieObject}{object created by "ieObject"}
\item{tree}{object of class"phylo"}
\item{branchWidth}{scaling factor for visualizing branch widths}
\item{branchCutoff}{number of standard deviations away from the mean rate. Used to indicate which branches should be colored.}
\item{labelSize}{size of ancestral node label}
\item{scaling}{scaling factor for visualizing node labels}
}
\value{
phylogenetic tree with mapped change and ancestral states
}
\description{
This function maps the results of an ie analysis onto a tree
}
\author{Jeroen Smaers}

