% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/ieRateDistr.r
\name{ieRateDistr}
\alias{ieRateDistr}
\title{Morphospace for a particular snapshot in evolutionaty time}
\usage{
ieRateDistr(Y, X, neutral)
}
\arguments{
\item{Y}{object created by "ieObject" for the Y variable}
\item{X}{object created by "ieObject" for the X variable}
\item{neutral}{number indicating the cut off to assign color to major axis residuals}
}
\value{
list of ratedistribution elements and branches for each quadrant
}
\description{
This function allocates the rate distribution between 2 variables to rate quadrants
}
\author{Jeroen Smaers}

