% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorphospaceEpoch.r
\name{MorphoSpaceEpoch}
\alias{MorphoSpaceEpoch}
\title{Function to calculate ancestral values at particular snapshots in evolutionary time}
\usage{
MorphoSpaceEpoch(X, Mya)
}
\arguments{
\item{X}{ieObject}
}
\value{
dataframe with rates for all branches and ancestral states for all nodes in the tree up to the indicated epoch
}
\description{
This function calculates ancestral values at a particular snapshot in evolutionare time
}

