% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBM.getRate.R
\name{mvBM.getRate}
\alias{mvBM.getRate}
\title{Lineage-specific rate estimation using multiple variance Brownian motion}
\usage{
mvBM.getRate(tree, tree_mvBM, branches, sigma2Distr)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{tree_mvBM}{an object of class "phylo". The rescaled tree from an mvBM procedure.}

\item{branches}{vector listing the branch numbers ('edge' numbers) for which the mvBM rate should be computed}

\item{sigma2Distr}{the MCMC distribution of sigma2 from an MCMC mvBM procedure}
}
\value{
mvBM rate estimate
}
\description{
Computes lineage-specific rates using mvBM
}
\examples{
see smaerslab.com/mvBM/ 
}
\references{
Smaers, Mongle & Kandler (2016) A multiple variance Brownian motion framework for estimating variable rates and inferring ancestral states. Biological Journal of the Linnean Society. 118 (1): 78-94.
}
