% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{primateData}
\alias{primateData}
\title{Isler et al. (2008) dataset of primate brain and body mass.}
\format{A data frame with 74 rows and 2 variables:
\describe{
  \item{Brain}{brain mass, in grams}
  \item{Body}{body mass, in grams}
  ...
}}
\source{
Isler K, Kirk EC, Miller JMA, Albrecht GA, Gelvin BR, Martin RD. 2008. Endocranial volumes of primate species: scaling analyses using a comprehensive and reliable data set. Journal of Human Evolution 55: 967-978.
}
\usage{
primateData
}
\description{
A dataset containing the brain and body mass of 74 primate species
}
\keyword{datasets}
