% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{primateTree}
\alias{primateTree}
\title{Molecular consensus phylogeny of primate species from the 10k trees project}
\format{A tree  146 primate species
  ...}
\source{
Arnold C, Matthews LJ, Nunn CL. 2010. The 10k Trees Website: A New Online Resource for Primate Phylogeny. Evolutionary Anthropology 19: 114-118.
}
\usage{
primateTree
}
\description{
An object of class "phylo"
}
\keyword{datasets}
