% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvoMorphoSpace.r
\name{EvoMorphoSpace}
\alias{EvoMorphoSpace}
\title{Evolutionary morphospace}
\usage{
EvoMorphoSpace(tree, X, Y, sequence, colors)
}
\arguments{
\item{tree}{an object of class "phylo".}

\item{X}{ieObject of variable 1}

\item{sequence}{a vector indicating the evolutionary snapshots to be considered}

\item{colors}{a vector indicating the colors of the data points along different evolutionary branches}

\item{X}{ieObject of variable 2}
}
\value{
dataframe with rates for all branches and ancestral states for all nodes in the tree up to the indicated epoch
}
\description{
This function computes a sequence of evolutionary snapshots of a bivariate morphospace
}

