% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEdges.r
\name{getEdges}
\alias{getEdges}
\title{List all edges descendant from a node}
\usage{
getEdges(tree, node)
}
\arguments{
\item{tree}{an object of class 'phylo'.}

\item{node}{number of interest.}
}
\value{
vector of edge numbers.
}
\description{
List all edges descendant from a node
}
