% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls.ancova.r
\name{gls.ancova}
\alias{gls.ancova}
\title{pGLS model comparison using phylogenetic analysis of covariance (gls.ancova)}
\usage{
gls.ancova(form, Sigma, ReducedModel, FullModel)
}
\arguments{
\item{form}{Model formula.}

\item{Sigma}{Matrix describing the covariance among data points. For the phylogenetic regression, this is based on a phylogeny (vcv(tree)).}

\item{ReducedModel}{The GLS model that includes the fewest indicator variables.}

\item{FullModel}{The GLS model that includes the fewest indicator variables.}
}
\value{
Anova table.
}
\description{
Computes F-ratio test of GLS models that include different sets of indicator variables.
}
\examples{
see https://smaerslab.com/software/ 
}
\references{
Smaers & Rohlf (2016) Testing species' deviations from allometric preductions using the phylogenetic regression. Evolution. 70 (5): 1145-1149.
}
