% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ieObject.r
\name{ieObject}
\alias{ieObject}
\title{Function to sort data according to ie output (following the ape tree structure)}
\usage{
ieObject(tree, data_extants, data_nodes)
}
\arguments{
\item{data_nodes}{named vector of ancestral states}
}
\value{
object with extant and nodal values sorted according to the ape tree structure
}
\description{
This function sorts extant and nodal values according the ape tree structure
}

