% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/iePlot.r
\name{iePlot}
\alias{iePlot}
\title{Function for mapping ie results onto a phylogenetic tree}
\usage{
iePlot(ieObject, tree, branchWidth, branchCutoff, labelSize, scaling)
}
\arguments{
\item{ieObject}{object comprising the output of an ie analysis}
}
\value{
phylogenetic tree with mapped change and ancestral states
}
\description{
This function maps the results of an ie analysis onto a tree
}

