% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieRateDistr.r
\name{ieRateDistr}
\alias{ieRateDistr}
\title{Function to allocate the rate distribution between 2 variables to rate quadrants}
\usage{
ieRateDistr(Y, X, neutral)
}
\arguments{
\item{ieObject}{object comprising the output of an ie analysis}
}
\value{
list of ratedistribution elements and branches for each quadrant
}
\description{
This function allocates the rate distribution between 2 variables to rate quadrants
}

