% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvBM.r
\name{mvBM}
\alias{mvBM}
\title{Multiple variance Brownian motion estimation}
\usage{
mvBM(data, tree, sigma2)
}
\arguments{
\item{data}{a vector of tip values for species; should be in the same order as tiplabels in the tree}

\item{tree}{an object of class "phylo".}

\item{sigma}{sig2 value from a Bayesian MCMC run using standard Brownian motion (e.g. ?anc.Bayes)}
}
\value{
dataframe with rescaled branch lengths (rBL) for all branches in the tree
}
\description{
Computes rescaled branch lengths using multiple variance Brownian motion as described in Smaers et al. (2016)
}
\examples{
tree<-pbtree(n=50)
x<-fastBM(tree,sig2=2) # simulate using fastBM
X<-anc.Bayes(tree,x,ngen=10000) # sample ancestral states
estimates<-colMeans(X[21:nrow(X),]) # get estimates, excluding burnin
sigma2<-mean(X[21:nrow(X),2]) # get sigma2, excluding burnin
mvBMresults<-mvBM(x,tree,sigma2) # calculate rescaled branch lengths using mvBM
tree_mvBM<-tree
tree_mvBM$edge.length<-mvBMresults$rBL # create new tree with rescaled branch lengths
X_mvBM<-anc.Bayes(tree_mvBM,x,ngen=10000)  # use new tree in Bayesian MCMC run
estimates_mvBM<-colMeans(X_mvBM[21:nrow(X_mvBM),]) # get estimates, excluding burnin
}
\references{
Smaers, Mongle & Kandler (2016) A multiple variance Brownian motion framework for estimating variable rates and inferring ancestral states. Biological Journal of the Linnean Society.
}

