% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pGLS.plotGrade.R
\name{pGLS.plotGrade}
\alias{pGLS.plotGrade}
\title{pGLS plot of a particular grade}
\usage{
pGLS.plotGrade(Yvar, Xvar, data, tree, group, ...)
}
\arguments{
\item{Yvar}{column name of the dependent variable}

\item{Xvar}{column name of the independent variable}

\item{data}{a vector of tip values for species; should be in the same order as tiplabels in the tree}

\item{tree}{an object of class "phylo".}

\item{group}{vector listing the tip labels for which the pGLS model should be plotted}
}
\value{
pGLS abline and points of the group of interest
}
\description{
Plots the pGLS model of a particular subset of a comparative dataset
}
\examples{
see https://smaerslab.com/software/ 
}
